requirements.txt
jax==0.5
flax==0.10.2
optax==0.2.4
orbax-checkpoint==0.11.1
hydra-core==1.3.2
Flask==3.1.0
requests==2.32.3